+++ Assets/Game/Building/Test2.asset
@@ -15,7 +15,7 @@ MonoBehaviour:
   DefaultGameObject: {fileID: 1508726255514112787, guid: 8ba1917497413bc4eb88f6b28139bb96,
     type: 3}
   Stages:
-  - objectName: "\u0417\u0430\u0431\u043E\u0440"
+  - ObjectName: "\u0410\u0431\u0443\u0431\u0430"
     Gameobject: {fileID: 9131482177403923620, guid: 9fd2646d1f1c356469e7fc4105ef4aae,
       type: 3}
     Requirements:
