+++ Assets/Scripts/BuildingRepair/ResourceRequirement.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 92dbf90fd3334ad6b512246d520f84c3
+timeCreated: 1740072743
