+++ Assets/Scripts/BuildingRepair/ResourceTransferButtonType.cs
@@ -1,0 +1,12 @@
+﻿namespace FlavorfulStory.BuildingRepair
+{
+    /// <summary> Типы операций передачи ресурсов. </summary>
+    public enum ResourceTransferButtonType
+    {
+        /// <summary> Добавление ресурса в ремонт. </summary>
+        Add,
+
+        /// <summary> Возврат ресурса в инвентарь. </summary>
+        Return
+    }
+}
