--- Assets/Scripts/ResourceTransferButtonType.cs
@@ -1,9 +1,0 @@
-﻿/// <summary> Типы операций передачи ресурсов. </summary>
-public enum ResourceTransferButtonType
-{
-    /// <summary> Добавление ресурса в ремонт. </summary>
-    Add,
-
-    /// <summary> Возврат ресурса в инвентарь. </summary>
-    Return
-}
